options.go: move the `Option` definitions to their own file
