+++ pipe/pipeline.go
@@ -44,10 +44,6 @@ type EnvVar struct {
 	Value string
 }
 
-type ContextValueFunc func(context.Context) (string, bool)
-
-type ContextValuesFunc func(context.Context) []EnvVar
-
 // Pipeline represents a Unix-like pipe that can include multiple
 // stages, including external processes but also and stages written in
 // Go.
@@ -84,104 +80,6 @@ func New(options ...Option) *Pipeline {
 	}
 
 	return p
-}
-
-// Option is a type alias for Pipeline functional options.
-type Option func(*Pipeline)
-
-// WithDir sets the default directory for running external commands.
-func WithDir(dir string) Option {
-	return func(p *Pipeline) {
-		p.env.Dir = dir
-	}
-}
-
-// WithStdin assigns stdin to the first command in the pipeline. The
-// caller retains ownership of stdin; the pipeline will not close it,
-// even if `Start()` returns an error.
-func WithStdin(stdin io.Reader) Option {
-	return func(p *Pipeline) {
-		p.stdin = Input(stdin)
-	}
-}
-
-// WithStdout assigns stdout to the last command in the pipeline. The
-// caller retains ownership of stdout; the pipeline will not close it,
-// even if `Start()` returns an error.
-func WithStdout(stdout io.Writer) Option {
-	return func(p *Pipeline) {
-		p.stdout = Output(stdout)
-	}
-}
-
-// WithStdoutCloser assigns stdout to the last command in the
-// pipeline, and closes stdout when the pipeline is done with it. The
-// pipeline is responsible for closing stdout even if `Start()` returns
-// an error.
-func WithStdoutCloser(stdout io.WriteCloser) Option {
-	return func(p *Pipeline) {
-		p.stdout = ClosingOutput(stdout)
-	}
-}
-
-// WithEnvVar appends an environment variable for the pipeline.
-func WithEnvVar(key, value string) Option {
-	return func(p *Pipeline) {
-		p.env.Vars = append(p.env.Vars, func(_ context.Context, vars []EnvVar) []EnvVar {
-			return append(vars, EnvVar{Key: key, Value: value})
-		})
-	}
-}
-
-// WithEnvVars appends several environment variable for the pipeline.
-func WithEnvVars(b []EnvVar) Option {
-	return func(p *Pipeline) {
-		p.env.Vars = append(p.env.Vars, func(_ context.Context, a []EnvVar) []EnvVar {
-			return append(a, b...)
-		})
-	}
-}
-
-// WithEnvVarFunc appends a context-based environment variable for the pipeline.
-func WithEnvVarFunc(key string, valueFunc ContextValueFunc) Option {
-	return func(p *Pipeline) {
-		p.env.Vars = append(p.env.Vars, func(ctx context.Context, vars []EnvVar) []EnvVar {
-			if val, ok := valueFunc(ctx); ok {
-				return append(vars, EnvVar{Key: key, Value: val})
-			}
-			return vars
-		})
-	}
-}
-
-// WithEnvVarsFunc appends several context-based environment variables for the pipeline.
-func WithEnvVarsFunc(valuesFunc ContextValuesFunc) Option {
-	return func(p *Pipeline) {
-		p.env.Vars = append(p.env.Vars, func(ctx context.Context, vars []EnvVar) []EnvVar {
-			return append(vars, valuesFunc(ctx)...)
-		})
-	}
-}
-
-// WithEventHandler sets a handler for the pipeline. Setting one will emit
-// and event for each process.
-func WithEventHandler(handler EventHandler) Option {
-	return func(p *Pipeline) {
-		p.eventHandler = handler
-	}
-}
-
-// WithStagePanicHandler sets a panic handler for the stages within a pipeline.
-// When a pipeline stage panics, the provided handler will be invoked, allowing
-// the client to handle the panic in whatever way they see fit.
-//
-// Note:
-//   - The client is responsible for deciding whether to recover from the panic or panicking again.
-//   - If a panic handler is not set, the panic will be propagated normally.
-func WithStagePanicHandler(ph StagePanicHandler) Option {
-	return func(p *Pipeline) {
-		p.panicHandler = ph
-	}
 }
 
 func (p *Pipeline) hasStarted() bool {
